[update]use double and some update

+++ Algorithms/simplebeamsearch.h
@@ -4,6 +4,7 @@
 #include <algorithm>
 #include <set>
 #include <queue>
+#include <cmath>
 #include "algorithmwrapper.h"
 
 class SimpleBeamSearch : public AlgorithmWrapper
@@ -14,7 +15,7 @@ public:
 
 private:
 
-    std::vector<int> beamSearch(int agent_index);
+    std::vector<double> beamSearch(int agent_index);
 
     const int max_depth = 5;
     const int max_width = 30;
